More work on class extension
